refactoring

+++ NeumorphTest/Workouts/WorkoutView+SubViews/DetailWorkoutView.swift
@@ -6,25 +6,26 @@
 //
 
 import SwiftUI
+import RealmSwift
 
 struct DetailWorkoutView: View {
     
     @EnvironmentObject var realmManager: RealmManager
+    @ObservedRealmObject var workout: Workout
+    @FocusState var isFocused: Bool?
     
     @State private var icon: String = ""
     @State var deleteMode = false
-    
-    @Binding var choosenExercises: [ChoosenExercise]
-    
+        
     var body: some View {
 
             ScrollView {
-                ForEach($choosenExercises.indices, id: \.self) {index in
+                ForEach(workout.choosenExercises, id: \.id) { choosenExercise in
                     ChoosenExerciseButtonView(
-                        choosenExercise: $choosenExercises[index],
+                        isFocused: _isFocused,
+                        choosenExercise: choosenExercise,
                         deleteMode: $deleteMode,
-                        action: {deleteChoosenExercise(index: index)}
-                    )
+                        action: {deleteChoosenExercise(choosenExerciseId: choosenExercise.id)})
                         .environmentObject(realmManager)
                 }
                 .toolbar {
@@ -36,6 +37,16 @@ struct DetailWorkoutView: View {
                                     .foregroundColor(.customRed)
                         }
                     }
+                    ToolbarItemGroup(placement: .keyboard) {
+                        HStack{
+                            Spacer()
+                            Button {
+                                isFocused = nil
+                            } label: {
+                                Text("Готово")
+                            }
+                        }
+                    }
                 }
                 .padding()
         }
@@ -43,19 +54,15 @@ struct DetailWorkoutView: View {
 }
 
 extension DetailWorkoutView {
-    private func deleteChoosenExercise(index: Int) {
-        let exerciseToDelete = choosenExercises[index]
-        choosenExercises.remove(at: index)
-        
-        exerciseToDelete.sets.forEach { sett in
-            realmManager.deleteSet(set: sett)
-        }
+    private func deleteChoosenExercise(choosenExerciseId: ObjectId) {
+        guard let exerciseToDelete = workout.choosenExercises.first(where: { $0.id == choosenExerciseId }) else {return}
+            realmManager.deleteAllSets(choosenExerciseId: exerciseToDelete.id)
             realmManager.deleteChoosenExercise(id: exerciseToDelete.id)
     }
 }
 
 struct DetailWorkoutView_Previews: PreviewProvider {
     static var previews: some View {
-        DetailWorkoutView(choosenExercises: .constant([]))
+        DetailWorkoutView(workout: Workout())
     }
 }
